mise a jour aeroport et ville

+++ src/main/java/tpVol/dao/DaoAeroport.java
@@ -2,6 +2,6 @@ package tpVol.dao;
 
 import tpVol.model.Aeroport;
 
-public interface DaoAeroport extends DaoGeneric<Aeroport, Integer>{
+public interface DaoAeroport extends DaoGeneric<Aeroport, Long>{
 
 }
