+++ src/main/java/tpVol/dao/DaoVille.java
@@ -2,6 +2,6 @@ package tpVol.dao;
 
 import tpVol.model.Ville;
 
-public interface DaoVille extends DaoGeneric<Ville, Integer>{
+public interface DaoVille extends DaoGeneric<Ville, Long>{
 
 }
